Remove A/B test
