+++ src/features/share/pages/ShareAppModal.native.test.tsx
@@ -1,10 +1,8 @@
 import React from 'react'
 
 import * as Share from 'features/share/helpers/shareApp'
-import { ShareAppModalType, ShareAppWordingVersion } from 'features/share/types'
+import { ShareAppModalType } from 'features/share/types'
 import { analytics } from 'libs/analytics'
-import { DEFAULT_REMOTE_CONFIG } from 'libs/firebase/remoteConfig/remoteConfig.constants'
-import * as useRemoteConfigContext from 'libs/firebase/remoteConfig/RemoteConfigProvider'
 import { fireEvent, render, screen, waitFor } from 'tests/utils'
 
 import { ShareAppModal } from './ShareAppModal'
@@ -13,30 +11,18 @@ const visible = true
 const hideModal = jest.fn()
 const shareApp = jest.spyOn(Share, 'shareApp').mockResolvedValue()
 
-const useRemoteConfigContextSpy = jest
-  .spyOn(useRemoteConfigContext, 'useRemoteConfigContext')
-  .mockReturnValue(DEFAULT_REMOTE_CONFIG)
-
 describe('ShareAppModal', () => {
-  it.each(['default', 'statistics', 'short'])(
-    'should match snapshot for %s wording',
-    (wordingVersion) => {
-      useRemoteConfigContextSpy.mockReturnValueOnce({
-        ...DEFAULT_REMOTE_CONFIG,
-        shareAppWordingVersion: wordingVersion as ShareAppWordingVersion,
-      })
+  it('should match snapshot', () => {
+    render(
+      <ShareAppModal
+        visible={visible}
+        hideModal={hideModal}
+        modalType={ShareAppModalType.NOT_ELIGIBLE}
+      />
+    )
 
-      render(
-        <ShareAppModal
-          visible={visible}
-          hideModal={hideModal}
-          modalType={ShareAppModalType.NOT_ELIGIBLE}
-        />
-      )
-
-      expect(screen).toMatchSnapshot()
-    }
-  )
+    expect(screen).toMatchSnapshot()
+  })
 
   it('should open native share modal when clicking on "Partager l’appli" button', async () => {
     render(
