+++ src/features/share/pages/ShareAppModal.tsx
@@ -3,16 +3,15 @@ import styled from 'styled-components/native'
 
 import { SHARE_APP_IMAGE_SOURCE } from 'features/share/components/shareAppImage'
 import { shareApp } from 'features/share/helpers/shareApp'
-import { shareAppModalInformations } from 'features/share/helpers/shareAppModalInformation'
 import { ShareAppModalType } from 'features/share/types'
 import { analytics } from 'libs/analytics'
-import { useRemoteConfigContext } from 'libs/firebase/remoteConfig/RemoteConfigProvider'
 import { ButtonQuaternaryBlack } from 'ui/components/buttons/ButtonQuaternaryBlack'
 import { ButtonWithLinearGradient } from 'ui/components/buttons/buttonWithLinearGradient/ButtonWithLinearGradient'
 import { MarketingModal } from 'ui/components/modals/MarketingModal'
 import { Share } from 'ui/svg/icons/BicolorShare'
 import { Invalidate } from 'ui/svg/icons/Invalidate'
-import { Spacer } from 'ui/theme'
+import { Spacer, Typo } from 'ui/theme'
+import { LINE_BREAK } from 'ui/theme/constants'
 
 type Props = {
   visible: boolean
@@ -21,8 +20,6 @@ type Props = {
 }
 
 export const ShareAppModal: FunctionComponent<Props> = ({ visible, hideModal, modalType }) => {
-  const { shareAppWordingVersion } = useRemoteConfigContext()
-
   const openShareAppModal = useCallback(() => {
     analytics.logShareApp({ type: modalType })
     hideModal()
@@ -36,15 +33,19 @@ export const ShareAppModal: FunctionComponent<Props> = ({ visible, hideModal, mo
     hideModal()
   }, [modalType, hideModal])
 
-  const { title, subtitleComponent } = shareAppModalInformations(shareAppWordingVersion)
-
   return (
     <MarketingModal
       visible={visible}
-      title={title}
+      title={'Fais tourner le bon plan\u00a0!'}
       imageSource={SHARE_APP_IMAGE_SOURCE}
       onBackdropPress={closeModal}>
-      {subtitleComponent}
+      <StyledBody>
+        <Typo.ButtonText>
+          35&nbsp;% des jeunes en France n’ont pas encore le pass Culture.
+        </Typo.ButtonText>
+        {LINE_BREAK}
+        Fais découvrir le pass à tes amis&nbsp;!
+      </StyledBody>
       <Spacer.Column numberOfSpaces={6} />
       <ButtonContainer>
         <ButtonWithLinearGradient
@@ -65,3 +66,6 @@ export const ShareAppModal: FunctionComponent<Props> = ({ visible, hideModal, mo
 }
 
 const ButtonContainer = styled.View({ width: '100%' })
+const StyledBody = styled(Typo.Body)({
+  textAlign: 'center',
+})
